style & refactor: 1. change the entry file of the module to be index.js 2. add some styling to the responses in a single review tile 3. factor out the helper functions for getting avg review to a seperate file
