+++ fec-incredibles/client/src/scss/reviews.scss
@@ -49,7 +49,13 @@
 .review-tile {
   display: flex;
   flex-flow: column nowrap;
-  margin: 4px;
+
+  padding: 20px 15px 10px 15px;
+  margin: 10px 0;
+
+  border: 1.5px outset #96a0c4;
+  border-radius: 10px;
+  box-shadow: 0 8px 16px -8px rgba(0,0,0,0.4);
 
   .review-tile-header {
     display: flex;
@@ -103,6 +109,17 @@
 
   .review-tile-response {
     margin: 8px 0;
+    padding: 8px 5px;
+    background-color: #8c888814;
+
+    b {
+      font-weight: bold;
+      margin: 5px 0;
+    }
+
+    p {
+      margin: 10px 0 5px 0;
+    }
   }
 
   .review-tile-footer {
